refactor: update references
